requirements.txt
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
boltons @ file:///C:/ci_311/boltons_1677729932371/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_7bbxxprxir/croot/certifi_1720453609121/work/certifi
cffi @ file:///C:/b/abs_78eb1_vq6z/croot/cffi_1714483206096/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama @ file:///C:/ci_311/colorama_1676422310965/work
conda @ file:///C:/b/abs_69e_za3rd5/croot/conda_1722004627938/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
cryptography @ file:///C:/b/abs_97lgskwerx/croot/cryptography_1714664736250/work
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
dj-database-url==2.2.0
Django==5.1.1
django-filter==24.3
django-heroku==0.3.1
django-taggit==6.0.0
djangorestframework==3.15.2
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
gunicorn==23.0.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_89suv9zwkz/croot/mamba-split_1714484512428/work/libmambapy
menuinst @ file:///C:/b/abs_76jwd47ck1/croot/menuinst_1723567618942/work
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pillow==10.4.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy @ file:///C:/ci_311/pluggy_1676422178143/work
psycopg==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
sqlparse==0.5.1
tqdm @ file:///C:/b/abs_a7h7n45ccq/croot/tqdm_1716395987133/work
truststore @ file:///C:/b/abs_55z7b3r045/croot/truststore_1695245455435/work
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
whitenoise==6.7.0
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
zstandard @ file:///C:/b/abs_e46tf0_91g/croot/zstandard_1714678303615/work
